pyproject.toml
[tool.poetry]
name = "realestate-connect"
version = "0.1.0"
description = "Real Estate connect backend"
authors = ["Eric Alfred <alfrederic371@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
appwrite = "^7.0.1"
pillow = "^11.0.0"
drf-spectacular = "^0.28.0"
channels = "^4.2.0"
redis = "^5.2.1"
websockets = "^14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
